instagram-service/requirements.txt
APScheduler==3.8.1
async-generator==1.10
attrs==21.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
chromedriver-autoinstaller==0.2.2
click==8.0.3
cryptography==36.0.0
Flask==2.0.2
Flask-APScheduler==1.12.2
Flask-Cors==3.0.10
h11==0.12.0
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
outcome==1.1.0
pycparser==2.21
pymongo==3.12.1
PyMySQL==1.0.2
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
Werkzeug==2.0.2
wsproto==1.0.0
zipp==3.6.0
